Fix first round after a server restart being displayed as number 0 (#14194)
